requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.11.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0

